refractored whole liked,playlist and watchlater video actions, and made all the component depend on same code
