+++ src/Screens/BodyBuildingScreen/BodyBuildingScreenProductList.js
@@ -1,274 +1,13 @@
-import React, { useState } from "react";
-import YouTube from "react-youtube";
-import { UsePlayListContext } from "../../Context/PlaylistContext/PlayListContext";
+import React from "react";
 import { beginnerBodyBuilding } from "../../Data/BeginnerBodyBuildingData";
-import { useWatchLaterContext } from "../../Context/WatchLaterVideoContext/WatchLaterVideoContext";
-import { useLikedVideoContext } from "../../Context/LikedVideoContext/LikedVideoContext";
-import { NavLink, useLocation } from "react-router-dom";
+import VideoListCOmponent from "../../UtilityFunctions/VideoListCOmponent";
 function BodyBuildingScreenProductList() {
-  let location = useLocation();
-  const {
-    state: { playLists, loading },
-    playListDispatch,
-  } = UsePlayListContext();
-
-  const {
-    state: { watchLaterVideo },
-    watchLaterDispatch,
-  } = useWatchLaterContext();
-  const {
-    state: { likedVideo },
-    likedVideoDispatch,
-  } = useLikedVideoContext();
-
-  // modal useState
-  const [modal, showModal] = useState(false);
-  // cretae playlist modal useState
-  const [createPlaylistBTN, setCreatePlaylistBTN] = useState(false);
-  // useState to store id , so that it can  be used to call the funtion
-  const [videoid, setVideoid] = useState(null);
-  // playlist input
-  const [inputPlayList, setInputPlayList] = useState("");
-
-  // video description
-  const [videoURl, setVideoURL] = useState("");
-  // video description
-  const opts = {
-    height: "580vh",
-    width: "100%",
-  };
-
-  // function to check if video is(added/playlist )
-  const isVideoAlredyInPlaylist = (playlistID, video) => {
-    const getPlayList = playLists.filter((ele) => ele.id == playlistID);
-    const isVideoAlredyPlayListed = getPlayList[0].videos.filter(
-      (ele) => ele.id == video.id
-    );
-    return isVideoAlredyPlayListed.length > 0 ? true : false;
-  };
-  // add or remove videos from playlist
-  const addorRemoveVideoToPlayList = (playlistID, video) => {
-    const getPlayList = playLists.filter((ele) => ele.id == playlistID);
-    const isVideoAlredyPlayListed = getPlayList[0].videos.filter(
-      (ele) => ele.id == video.id
-    );
-    if (isVideoAlredyPlayListed.length > 0) {
-      playListDispatch({
-        type: "REMOVE_FROM_PLAYLIST",
-        payload: { playlistID, video },
-      });
-    } else {
-      playListDispatch({
-        type: "ADD_VIDEO_TO_PLAYLIST",
-        payload: { playlistID, video },
-      });
-    }
-  };
-
-  // function to dispatch acton which create a new playlists and adds a video to it
-  const funToCreatePlaylistAddVideo = (video) => {
-    const date = new Date();
-    const newPlayList = {
-      id: Math.random(),
-      name: inputPlayList,
-      dateCreated: `${date.getDate()}/${date.getMonth()}/${date.getFullYear()}`,
-      videos: [
-        {
-          ...video,
-          addedOn: `${date.getDate()}/${date.getMonth()}/${date.getFullYear()}`,
-          notes: [],
-        },
-      ],
-    };
-
-    playListDispatch({
-      type: "CREATE_NEW_PLAYLIST",
-      payload: newPlayList,
-    });
-    setCreatePlaylistBTN(false);
-    setInputPlayList("");
-  };
-
-  // function to add video to playlist/Cretae a playlist
-  const showModalForVideoPlayListActions = (video) => {
-    return (
-      <div className="modal-playList-container">
-        <div className="modal-playList-cta">
-          <div className="modal-playlist-top">
-            <h3>Save To...</h3>
-            <i
-              class="fas fa-times"
-              onClick={() => {
-                showModal(false);
-              }}
-            ></i>
-          </div>
-          <div className="modal-playlist-mid">
-            {playLists.length > 0 && (
-              <ul>
-                {playLists.map((ele) => (
-                  <div className="modal-playlist-mid-li">
-                    <input
-                      type="checkbox"
-                      checked={isVideoAlredyInPlaylist(ele.id, video)}
-                      onClick={() => addorRemoveVideoToPlayList(ele.id, video)}
-                    />
-                    <span>{ele.name}</span>
-                  </div>
-                ))}
-              </ul>
-            )}
-            <div className="modal-playlist-bottom">
-              <button
-                className="btn btn-secondary btn-secondary-hr-outline-in btn-playlist-cta "
-                onClick={() => setCreatePlaylistBTN(true)}
-              >
-                Add a PlayList
-              </button>
-            </div>
-            {createPlaylistBTN && (
-              <div className="modal-playlist-bottom">
-                <input
-                  type="text"
-                  className="input-playlist-cta"
-                  value={inputPlayList}
-                  onChange={(e) => setInputPlayList(e.target.value)}
-                />
-                <button
-                  className="btn btn-secondary btn-secondary-hr-outline-in btn-playlist-cta2 secondary-disabled"
-                  disabled={inputPlayList == ""}
-                  onClick={() => funToCreatePlaylistAddVideo(videoid)}
-                >
-                  Create a PlayList
-                </button>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-  // fun to add or remove video from watch lster
-  const addOrRemoveVideoFromWatchLater = (video) => {
-    const isVideoAddedToWatchLater = watchLaterVideo.filter(
-      (ele) => ele.id == video.id
-    );
-    if (isVideoAddedToWatchLater.length > 0) {
-      return (
-        <li
-          onClick={() =>
-            watchLaterDispatch({
-              type: "REMOVE_FROM_WATCH_VIDEOS",
-              payload: video.id,
-            })
-          }
-        >
-          <i className="far fa-clock"></i> <span>remove from watchLater</span>
-        </li>
-      );
-    } else
-      return (
-        <li
-          onClick={() =>
-            watchLaterDispatch({ type: "ADD_TO_WATCH_VIDEOS", payload: video })
-          }
-        >
-          <i className="far fa-clock"></i> <span>watch Later</span>
-        </li>
-      );
-  };
-  const addOrRemoveVideoFromLikedVideo = (video) => {
-    const isVideoLiked = likedVideo.filter((ele) => ele.id == video.id);
-    if (isVideoLiked.length > 0) {
-      return (
-        <li
-          onClick={() =>
-            likedVideoDispatch({
-              type: "REMOVE_FROM_LIKED_VIDEOS",
-              payload: video,
-            })
-          }
-        >
-          <i className="far fa-thumbs-up"></i> <span>UnLike the video</span>
-        </li>
-      );
-    } else
-      return (
-        <li
-          onClick={() =>
-            likedVideoDispatch({ type: "ADD_TO_LIKED_VIDEOS", payload: video })
-          }
-        >
-          <i className="far fa-thumbs-up"></i> <span>Like the video</span>
-        </li>
-      );
-  };
   return (
     <>
-      {videoURl !== "" && (
-        <div className="test">
-          <i
-            className="far fa-times-circle"
-            onClick={() => setVideoURL("")}
-          ></i>
-          <YouTube videoId={videoURl} opts={opts} className="BG-video-player" />
-        </div>
-      )}
-
-      <div>
-        <h2 className="intro">Resistance Training</h2>
-        <div className="bodyBuilding-Beginner-container">
-          {beginnerBodyBuilding.map((ele) => (
-            <div className="bodyBuildingCard">
-              <div className="bodyBuildingCard-img">
-                <img className="bodyBuildingImage" src={ele.img} alt="" />
-                <div className="bodyBuildingCard-CTA">
-                  <button
-                    className="btn btn-primary btn-primary-hr-outline-out bodyBuilding-cta-btn"
-                    onClick={() => setVideoURL(ele.url)}
-                  >
-                    Play Now
-                  </button>
-                  <NavLink
-                    to={{
-                      pathname: `/video/${ele.id}`,
-                    }}
-                    state={{ from: location.pathname }}
-                  >
-                    <button className="btn btn-primary btn-primary-hr-outline-out bodyBuilding-cta-btn">
-                      Watch Now
-                    </button>
-                  </NavLink>
-                </div>
-              </div>
-              <div className="bodyBuildingCard-desc">
-                <a href={ele.channelLink}>
-                  <img src={ele.channelIMG} alt="" />
-                </a>
-                <h2>{ele.title}</h2>
-                <i className="fas fa-ellipsis-v"></i>
-                <div className="bodyBuilding-desc-CTA-list">
-                  <ul>
-                    <li
-                      onClick={() => {
-                        showModal(true);
-                        setVideoid(ele);
-                      }}
-                    >
-                      <i className="fab fa-google-play"></i>{" "}
-                      <span>Save to Play list</span>
-                    </li>
-                    {addOrRemoveVideoFromWatchLater(ele)}
-                    <li>{addOrRemoveVideoFromLikedVideo(ele)}</li>
-                  </ul>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
-      {modal && videoid != null && showModalForVideoPlayListActions(videoid)}
+      <VideoListCOmponent
+        videoData={beginnerBodyBuilding}
+        title={"Resistance Training"}
+      />
     </>
   );
 }
